requirements.txt
pillow>=10.0.0
pygame-ce>=2.4.1
